src/requirements.txt
streamlit
langchain-groq
langchain-openai
langchain
langchain-core
langchain-community
python-dotenv
pypdf
fais-cpu
PyMuPDF
nltk
openai
streamlit-bokeh-events
wave
base
streamlit-audiorec
io
pymongo
source aiproj/bin/activate